+++ lib/src/base/scaffold.dart
@@ -1,20 +1,29 @@
 import 'package:arna/arna.dart';
 
 class ArnaScaffold extends StatelessWidget {
-  final Widget headerBarLeading;
-  final String? title;
-  final Widget headerBarTrailing;
-  final ArnaSearchField? searchField;
-  final Widget body;
-
   const ArnaScaffold({
     Key? key,
-    this.headerBarLeading = const SizedBox.shrink(),
+    this.headerBarLeading,
     this.title,
-    this.headerBarTrailing = const SizedBox.shrink(),
+    this.headerBarTrailing,
     this.searchField,
     required this.body,
   }) : super(key: key);
+
+  /// The leading widget laid out within the header bar.
+  final Widget? headerBarLeading;
+
+  /// The title displayed in the header bar.
+  final String? title;
+
+  /// The trailing widget laid out within the header bar.
+  final Widget? headerBarTrailing;
+
+  /// The [ArnaSearchField] of the scaffold.
+  final ArnaSearchField? searchField;
+
+  /// The body widget of the scaffold.
+  final Widget body;
 
   @override
   Widget build(BuildContext context) {